requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
yandex-cloud-ml-sdk==0.1.2
openai>=1.0.0
google-generativeai>=0.3.0